Dashboardコンポーネントでfirebaseから他ユーザ情報を取得する処理に変更
