+++ throw-money-app/src/reducks/store/initialState.js
@@ -6,7 +6,7 @@ const initialState = {
     uid: '',
     userName: '',
     remainMoney: '',
-    otherUsersInfo: [],
+    //otherUsersInfo: [],
   },
 }
