+++ throw-money-app/src/reducks/users/actions.js
@@ -24,7 +24,7 @@ export const loginAction = (userState) => {
       uid: userState.uid,
       userName: userState.userName,
       remainMoney: userState.remainMoney,
-      otherUsersInfo: userState.otherUsersInfo,
+      //otherUsersInfo: userState.otherUsersInfo,
     },
   }
 }
